Refactored EventFragment End of the day commit
