+++ .idea/misc.xml
@@ -8,6 +8,7 @@
         <entry key="..\:/Users/fakgoz/Desktop/School/Projects/Cmpe451/2021SpringGroup4/app/src/main/res/layout/content_login.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/fakgoz/Desktop/School/Projects/Cmpe451/2021SpringGroup4/app/src/main/res/layout/fragment_dashboard.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/fakgoz/Desktop/School/Projects/Cmpe451/2021SpringGroup4/app/src/main/res/layout/fragment_first.xml" value="0.36666666666666664" />
+        <entry key="..\:/Users/fakgoz/Desktop/School/Projects/Cmpe451/2021SpringGroup4/app/src/main/res/layout/fragment_home.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/fakgoz/Desktop/School/Projects/Cmpe451/2021SpringGroup4/app/src/main/res/layout/fragment_lg.xml" value="0.18414387031408308" />
         <entry key="..\:/Users/fakgoz/Desktop/School/Projects/Cmpe451/2021SpringGroup4/app/src/main/res/layout/fragment_login.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/fakgoz/Desktop/School/Projects/Cmpe451/2021SpringGroup4/app/src/main/res/layout/fragment_profile.xml" value="0.36666666666666664" />
