+++ app/src/main/java/com/example/sportshub/ui/event/EventFragment.kt
@@ -1,4 +1,4 @@
-package com.example.sportshub.ui.home
+package com.example.sportshub.ui.event
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,13 +8,13 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import com.example.sportshub.R
-import com.example.sportshub.databinding.FragmentHomeBinding
+import com.example.sportshub.databinding.FragmentEventBinding
 
-class HomeFragment : Fragment() {
 
-    private lateinit var homeViewModel: HomeViewModel
-    private var _binding: FragmentHomeBinding? = null
+class EventFragment : Fragment() {
+
+    private lateinit var eventViewModel: EventViewModel
+    private var _binding: FragmentEventBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -24,17 +24,14 @@ class HomeFragment : Fragment() {
             inflater: LayoutInflater,
             container: ViewGroup?,
             savedInstanceState: Bundle?
-    ): View? {
-        homeViewModel =
-                ViewModelProvider(this).get(HomeViewModel::class.java)
+    ): View {
+        eventViewModel =
+                ViewModelProvider(this).get(EventViewModel::class.java)
 
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentEventBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textHome
-        homeViewModel.text.observe(viewLifecycleOwner, Observer {
-            textView.text = it
-        })
+
         return root
     }
