+++ app/src/main/java/com/example/sportshub/ui/event/EventViewModel.kt
@@ -1,0 +1,9 @@
+package com.example.sportshub.ui.event
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class EventViewModel : ViewModel() {
+    //TODO Implement ViewModel
+}
