+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
     <string name="app_name">SportsHub</string>
-    <string name="title_home">Home</string>
+    <string name="title_event">Event</string>
     <string name="title_dashboard">Dashboard</string>
     <string name="title_profile">Profile</string>
 </resources>
